requirements.txt
Flask>=2.2,<3
Django>=4.2,<5.0
Pillow>=10.0.0
